perf: base spline mise en cache, et retrait d'une précision mixte contre-productive
Lot A du plan de reprise, constats A4 et A5. Les deux ont été MESURÉS avant
d'être touchés.

A4 — gradient_metal.py construisait sa base spline par `spline_knot_count`
objets CubicSpline distincts, un par vecteur unité, À CHAQUE ÉVALUATION du
gradient, alors que le résultat est identique tant que les nœuds ne bougent pas :

    5 noeuds : 1,0 ms      12 noeuds : 2,4 ms
    8 noeuds : 1,6 ms      20 noeuds : 3,9 ms

soit ~3,4 s sur une optimisation de 2000 évaluations. Branché sur
SplineBasisCache, qui fait déjà exactement ce calcul et le mémorise :
1974 us -> 8,2 us, x241, pour un résultat rigoureusement identique (écart nul).

SplineBasisCache reçoit un paramètre `extrapolate`, intégré à sa clé : le
chemin metal a besoin de 0 hors du domaine des nœuds, les autres appelants d'une
extrapolation. Les deux variantes coexistent sans se mélanger — plutôt que
d'écrire un second cache.

A5 — la politique « précision mixte f32/c64 pour le débit SIMD » ne tient pas.
Mesure sur compute_TMM_generic, empilement de 12 couches, meilleur temps sur
5 passes de 5000 appels :

    c128/f64 : 0,9045 us/appel   erreur R = 0        (référence : oracle TMM)
    c64/f32  : 0,9144 us/appel   erreur R = 2,7e-08

Soit 1,1 % PLUS LENT, pour huit ordres de grandeur de précision perdus. LLVM ne
vectorise pas mieux le complex64 sur cette boucle, et les conversions coûtent
plus qu'elles ne rapportent.

Ces fonctions alimentent 20 sites d'appel, dont certus_opt_tmm.py:507 et les
workers de design. Une erreur de 2,7e-08 sur R passe sous le bruit d'un
spectrophotomètre, mais elle est désastreuse pour les gradients par différences
finies : avec un pas h = 1e-6, elle donne ~3 % d'erreur sur la dérivée.

Retour en double précision. Aucun test numérique n'a bougé — seules quatre
assertions portant directement sur le type ont dû suivre, ce qui confirme que
rien ne dépendait de la précision réduite.

+++ tests/unit/test_certus_core.py
@@ -217,13 +217,17 @@ class TestPrecisionConfig:
 
     def test_get_float_dtype(self):
         """Test get_float_dtype."""
+        # Double precision. La politique « precision mixte f32/c64 pour le debit
+        # SIMD » a ete mesuree le 2026-08-02 : 1,1 % PLUS LENTE que la double
+        # precision, pour 8 ordres de grandeur de precision perdus (erreur R de
+        # 2,7e-08 contre 0 sur compute_TMM_generic). Voir certus_core.py.
         dtype = get_float_dtype()
-        assert dtype == np.float32
+        assert dtype == np.float64
 
     def test_get_complex_dtype(self):
         """Test get_complex_dtype."""
         dtype = get_complex_dtype()
-        assert dtype == np.complex64
+        assert dtype == np.complex128
 
 
 class TestLoggingSystem:
